refactor: voucher domain object를 반환하는 service 메서드를 dto를 반환하게 변경
