+++ src/main/java/org/weekly/weekly/ui/reader/ConsoleWrap.java
@@ -3,7 +3,6 @@ package org.weekly.weekly.ui.reader;
 import org.springframework.stereotype.Component;
 
 import java.io.Console;
-import java.io.IOException;
 
 @Component
 public class ConsoleWrap implements CommandReader{
