+++ src/main/java/org/weekly/weekly/voucher/domain/Voucher.java
@@ -27,6 +27,22 @@ public class Voucher {
         return voucherId;
     }
 
+    public long getAmount() {
+        return amount;
+    }
+
+    public LocalDate getRegistrationDate() {
+        return registrationDate;
+    }
+
+    public LocalDate getExpirationDate() {
+        return expirationDate;
+    }
+
+    public Discount getDiscount() {
+        return discount;
+    }
+
     public long applyDiscount(long beforeAmount) {
         long afterAmount = discount.applyDiscount(beforeAmount, amount);
