+++ src/main/java/org/weekly/weekly/voucher/dto/Response.java
@@ -1,4 +1,4 @@
-package org.weekly.weekly.voucher.model;
+package org.weekly.weekly.voucher.dto;
 
 public interface Response<T> {
     String getResult();
